third_party/blink/web_tests/virtual/focusless-spat-nav/fast/spatial-navigation/focusless/snav-focusless-form-state-on-click.html

<!DOCTYPE html>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/page/spatial_navigation.mojom.js"></script>
<script src="../../../../../fast/spatial-navigation/resources/mock-snav-service.js"></script>
<script src="../../../../../fast/spatial-navigation/resources/snav-testharness.js"></script>

<form>
<input type="text" id="first"></input>
<input type="text" id="second"></input>
</form>

<script>
  function notCalled() {
    assert_false(true, "Should not be called");
  };
  mockSnavService.callback = notCalled;

  promise_test(async () => {
    if (accessibilityController) {
      accessibilityController.reset();
    }

    let first = document.getElementById("first");
    let second = document.getElementById("second");

    await snav.rAF();

    snav.triggerMove('Down');

    await snavCallback();
    assert_equals(window.internals.interestedElement,
                  first,
                  "'first' should be interested.");
    assert_not_equals(document.activeElement,
                      first,
                      "'first' should not yet be focused.");
    assert_false(mockSnavService.canExitFocus,
                 "Nothing focused, should not be able to exit focus.");
    assert_true(mockSnavService.canSelectElement,
                "Should be able to select element.");
    assert_false(mockSnavService.hasNextFormElement,
                 "Should be able to move to next form element.");

    eventSender.keyDown('Enter');
    await snavCallback();
    assert_true(mockSnavService.canExitFocus,
                "Should be able to exit focus.");
    assert_true(mockSnavService.canSelectElement,
                "Should be able to select element.");
    assert_true(mockSnavService.hasNextFormElement,
                "Should be able to move to next form element.");
    assert_equals(window.internals.interestedElement,
                  first,
                  "'first' should be interested.");
    assert_equals(document.activeElement,
                  first,
                  "'first' should be focused.");
  }, 'Spat Nav state updates correctly when clicking on form elements.');
</script>
